ZALANDO CODES DISCORD BOT GENERATOR/requirements.txt
discord.py==1.7.3
selenium==3.141.0
